+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ### Changed
 
 - Slight change to `FlattenedCE` and `unflatten` to ensure that basic functionality remains intact. [#505]
+- Fixed small issue in `benchmark` function.
 
 ## Version [1.4.1] - 2024-12-19